+++ src/render.rs
@@ -4,14 +4,142 @@ use crate::prelude::*;
 
 pub fn render(output: &wgpu::SurfaceTexture, program_data: &mut ProgramData) {
 	let output_view = output.texture.create_view(&wgpu::TextureViewDescriptor::default());
-	let encoder_descriptor = wgpu::CommandEncoderDescriptor {label: Some("Render Encoder")};
+	let encoder_descriptor = wgpu::CommandEncoderDescriptor {label: None};
 	let mut encoder = program_data.render_context.device.create_command_encoder(&encoder_descriptor);
+	
+	let frustum_planes = get_frustum_planes(&program_data.camera_data, program_data.render_context.aspect_ratio);
+	let visible_models_list = get_visible_models(
+		&program_data.example_model_instance_datas,
+		program_data.render_assets.example_models.bounding_radius,
+		frustum_planes
+	);
+	
+	update_gpu_buffers(program_data, &visible_models_list);
 	
 	render_shadow_caster_pipeline(program_data, &mut encoder);
 	render_models_pipeline(program_data, &mut encoder, &output_view);
-	render_skybox_pipeline(program_data, &mut encoder, &output_view); /// it's better to have this at the end so that only the necessary pixels are rendered
+	render_skybox_pipeline(program_data, &mut encoder, &output_view); // HELP: it's better to have this at the end so that only the necessary pixels are rendered
 	
 	program_data.render_context.command_queue.submit(std::iter::once(encoder.finish()));
+}
+
+
+
+// this is an implementation of frustum culling based on: https://learnopengl.com/Guest-Articles/2021/Scene/Frustum-Culling
+pub fn get_visible_models(instance_datas: &[InstanceData], bounding_radius: f32, frustum_planes: [(glam::Vec3, f32); 5]) -> Vec<usize> {
+	let mut output = vec!();
+	for (i, instance) in instance_datas.iter().enumerate() {
+		if model_is_visible(instance.pos, bounding_radius, &frustum_planes) {
+			output.push(i);
+		}
+	}
+	output
+}
+
+pub fn get_frustum_planes(camera: &CameraData, aspect_ratio: f32) -> [(glam::Vec3, f32); 5] {
+	let forward = glam::Vec3::new(
+		camera.rot_xz.cos() * camera.rot_y.cos(),
+		camera.rot_y.sin(),
+		camera.rot_xz.sin() * camera.rot_y.cos(),
+	).normalize();
+	
+	//let near_plane = { // there's not really any point in checking this plane but you still can if you just want to
+	//	let normal = forward;
+	//	let point_on_plane = camera.pos + forward * camera.near;
+	//	let dist = point_on_plane.dot(forward);
+	//	(normal, dist)
+	//};
+	
+	let far_plane = { // you might be able to remove this one too
+		let normal = -forward;
+		let point_on_plane = camera.pos + forward * camera.far;
+		let dist = point_on_plane.dot(normal);
+		(normal, dist)
+	};
+	
+	let up_dir = glam::Vec3::Y;
+	let right_dir = forward.cross(up_dir).normalize();
+	let up_dir = forward.cross(-right_dir).normalize();
+	const FOV_MULT: f32 = 0.9; // lower this to see the culling work
+	let half_height = (camera.fov_radians * FOV_MULT / 2.0).tan() * camera.near;
+	let half_width = half_height * aspect_ratio;
+	let near_plane_center = camera.pos + forward * camera.near;
+	
+	let left_plane = {
+		let point_on_plane = near_plane_center - right_dir * half_width;
+		let normal = -up_dir.cross(point_on_plane - camera.pos).normalize();
+		let dist = point_on_plane.dot(normal);
+		(normal, dist)
+	};
+	
+	let right_plane = {
+		let point_on_plane = near_plane_center + right_dir * half_width;
+		let normal = up_dir.cross(point_on_plane - camera.pos).normalize();
+		let dist = point_on_plane.dot(normal);
+		(normal, dist)
+	};
+	
+	let top_plane = {
+		let point_on_plane = near_plane_center + up_dir * half_height;
+		let normal = -right_dir.cross(point_on_plane - camera.pos).normalize();
+		let dist = point_on_plane.dot(normal);
+		(normal, dist)
+	};
+	
+	let bottom_plane = {
+		let point_on_plane = near_plane_center - up_dir * half_height;
+		let normal = right_dir.cross(point_on_plane - camera.pos).normalize();
+		let dist = point_on_plane.dot(normal);
+		(normal, dist)
+	};
+	
+	[far_plane, left_plane, right_plane, top_plane, bottom_plane]
+}
+
+pub fn model_is_visible(pos: glam::Vec3, bounding_radius: f32, frustum_planes: &[(glam::Vec3, f32); 5]) -> bool {
+	is_sphere_past_plane(pos, bounding_radius, frustum_planes[0].0, frustum_planes[0].1)
+		&& is_sphere_past_plane(pos, bounding_radius, frustum_planes[1].0, frustum_planes[1].1)
+		&& is_sphere_past_plane(pos, bounding_radius, frustum_planes[2].0, frustum_planes[2].1)
+		&& is_sphere_past_plane(pos, bounding_radius, frustum_planes[3].0, frustum_planes[3].1)
+		&& is_sphere_past_plane(pos, bounding_radius, frustum_planes[4].0, frustum_planes[4].1)
+}
+
+pub fn is_sphere_past_plane(pos: glam::Vec3, radius: f32, plane_normal: glam::Vec3, plane_dist: f32) -> bool {
+	plane_normal.dot(pos) > plane_dist - radius
+}
+
+
+
+pub fn update_gpu_buffers(program_data: &mut ProgramData, visible_models: &[usize]) {
+	
+	// culled_instances_buffer
+	let mut new_model_instances_data = Vec::with_capacity(visible_models.len());
+	for index in visible_models {
+		new_model_instances_data.push(program_data.example_model_instance_datas[*index].to_raw())
+	}
+	program_data.render_context.command_queue.write_buffer(
+		&program_data.render_assets.example_models.culled_instances_buffer,
+		0,
+		bytemuck::cast_slice(&new_model_instances_data),
+	);
+	program_data.render_assets.example_models.culled_instances_count = visible_models.len() as u32;
+	
+	// camera.buffer
+	let camera_gpu_data = program_data.camera_data.build_gpu_data(program_data.render_context.aspect_ratio);
+	program_data.render_context.command_queue.write_buffer(
+		&program_data.render_assets.camera.buffer,
+		0,
+		bytemuck::cast_slice(&camera_gpu_data),
+	);
+	
+	// shadow_caster.proj_mat_buffer
+	let shadow_caster_gpu_data = program_data.shadow_caster_data.build_gpu_data(program_data.camera_data.pos);
+	program_data.render_context.command_queue.write_buffer(
+		&program_data.render_assets.shadow_caster.proj_mat_buffer,
+		0,
+		bytemuck::cast_slice(&shadow_caster_gpu_data),
+	);
+	
 }
 
 
@@ -42,10 +170,9 @@ pub fn render_shadow_caster_pipeline(program_data: &ProgramData, encoder: &mut w
 	
 	for mesh in &render_assets.example_models.meshes {
 		shadow_caster_pass_handle.set_vertex_buffer(0, mesh.basic_vertex_buffer.slice(..));
-		shadow_caster_pass_handle.set_vertex_buffer(1, mesh.extended_vertex_buffer.slice(..)); // TODO: remove this line
-		shadow_caster_pass_handle.set_vertex_buffer(2, render_assets.example_models.instances_buffer.slice(..));
+		shadow_caster_pass_handle.set_vertex_buffer(1, render_assets.example_models.total_instances_buffer.slice(..));
 		shadow_caster_pass_handle.set_index_buffer(mesh.index_buffer.slice(..), wgpu::IndexFormat::Uint32);
-		shadow_caster_pass_handle.draw_indexed(0..mesh.index_count, 0, 0..render_assets.example_models.instances_count);
+		shadow_caster_pass_handle.draw_indexed(0..mesh.index_count, 0, 0..render_assets.example_models.total_instances_count);
 	}
 	
 }
@@ -91,9 +218,9 @@ pub fn render_models_pipeline(program_data: &ProgramData, encoder: &mut wgpu::Co
 		models_pass_handle.set_bind_group(1, &program_data.render_bindings.example_models_bind_1s[i], &[]);
 		models_pass_handle.set_vertex_buffer(0, mesh.basic_vertex_buffer.slice(..));
 		models_pass_handle.set_vertex_buffer(1, mesh.extended_vertex_buffer.slice(..));
-		models_pass_handle.set_vertex_buffer(2, render_assets.example_models.instances_buffer.slice(..));
+		models_pass_handle.set_vertex_buffer(2, render_assets.example_models.culled_instances_buffer.slice(..));
 		models_pass_handle.set_index_buffer(mesh.index_buffer.slice(..), wgpu::IndexFormat::Uint32);
-		models_pass_handle.draw_indexed(0..mesh.index_count, 0, 0..render_assets.example_models.instances_count);
+		models_pass_handle.draw_indexed(0..mesh.index_count, 0, 0..render_assets.example_models.culled_instances_count);
 	}
 	
 }