+++ changelog.md
@@ -3,6 +3,12 @@
 <br>
 <br>
 
+- 0.1.1 (24/05/20)
+  - Added frustum culling
+  - Removed extended vertex data from shadowmap rendering
+
+<br>
+
 - **0.1.0** (24/05/12)
   - Added changelog
   - Finished shadows